refactor: move channels component
