+++ app/channels/[broadcaster]/page.tsx
@@ -2,8 +2,8 @@
 
 import { useEffect, useState } from "react";
 
-import Chat from "@/components/channels/chat";
-import VideoPlayer from "@/components/channels/video-player";
+import Chat from "@/app/channels/[broadcaster]/components/chat";
+import VideoPlayer from "@/app/channels/[broadcaster]/components/video-player";
 import { useToast } from "@/components/ui/use-toast";
 import { SignalSlashIcon } from "@heroicons/react/24/solid";
 import { useSocket } from "@/app/components/socket-provider";
