+++ app/channels/components/channel-nav-item.tsx
